Install GTK3/GTK4
