+++ po/POTFILES.in
@@ -1,5 +1,5 @@
-# intltool-update -p -g folder_i18n
+# intltool-update -p -g folder-color
 # also done automatically with setup.py build or install
 [encoding: UTF-8]
-folder_path/folder-color.py
-
+extension/GKT3.py
+extension/GKT4.py
